+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-02-02
+## [Unreleased] - 2023-03-02
 ### Added
 - The readme file.
 - This change log.
 - The initial files.
+### Changed
+- Summon-Spiritual-Weapon now uses a mule (SpellMule) for its weapon name and URL.
+- Summon-Spiritual-Weapon now uses presents weapon and direction prompts.